Added security and swagger docs and modified to constructor injection

+++ src/main/java/com/products/productDetails/DTO/ResponseObject/ProductDetailsRO.java
@@ -12,7 +12,7 @@ public class ProductDetailsRO {
 
     private String productBrand;
 
-    private int productPrice;
+    private long productPrice;
 
     private int stockQuantity;
 }
